test(e2e): restore scenario logic to test file
